Rename `app` -> `futarchy-ts`
