+++ futarchy-ts/src/AutocratClient.ts
