+++ futarchy-ts/src/ConditionalVaultClient.ts
