+++ futarchy-ts/src/constants.ts
