+++ futarchy-ts/src/idl/amm.json
