+++ futarchy-ts/src/idl/autocrat.json
