+++ futarchy-ts/src/idl/autocrat_migrator.json
