+++ futarchy-ts/src/idl/autocrat_v0.json
