+++ futarchy-ts/src/index.ts
