+++ futarchy-ts/src/types/amm.ts
