+++ futarchy-ts/src/types/autocrat.ts
