+++ futarchy-ts/src/types/conditional_vault.ts
