+++ futarchy-ts/src/types/index.ts
