+++ futarchy-ts/src/utils/priceMath.ts
