+++ tests/amm.ts
@@ -17,13 +17,13 @@ import {
   getMint,
 } from "spl-token-bankrun";
 
-import { getAmmAddr, sleep } from "../app/src/utils";
+import { getAmmAddr, sleep } from "../futarchy-ts/src/utils";
 import { Keypair, PublicKey } from "@solana/web3.js";
 import { assert } from "chai";
-import { AmmClient } from "../app/src/AmmClient";
+import { AmmClient } from "../futarchy-ts/src/AmmClient";
 import { expectError, fastForward } from "./utils/utils";
-import { PriceMath } from "../app/src/utils/priceMath";
-import { getATA, getAmmLpMintAddr } from "../app/src/utils/pda";
+import { PriceMath } from "../futarchy-ts/src/utils/priceMath";
+import { getATA, getAmmLpMintAddr } from "../futarchy-ts/src/utils/pda";
 
 const META_DECIMALS = 9;
 const USDC_DECIMALS = 6;
