+++ tests/autocrat.ts
@@ -39,9 +39,9 @@ import {
   getAmmAddr,
   getAmmLpMintAddr,
   getVaultAddr,
-} from "../app/src";
-import { PriceMath } from "../app/src/utils/priceMath";
-import { AutocratClient } from "../app/src/AutocratClient";
+} from "../futarchy-ts/src";
+import { PriceMath } from "../futarchy-ts/src/utils/priceMath";
+import { AutocratClient } from "../futarchy-ts/src/AutocratClient";
 import {
   ComputeBudgetInstruction,
   ComputeBudgetProgram,
@@ -49,7 +49,7 @@ import {
   Transaction,
   TransactionInstruction,
 } from "@solana/web3.js";
-import { ConditionalVaultClient } from "../app/src/ConditionalVaultClient";
+import { ConditionalVaultClient } from "../futarchy-ts/src/ConditionalVaultClient";
 
 const AutocratIDL: Autocrat = require("../target/idl/autocrat.json");
 const ConditionalVaultIDL: ConditionalVault = require("../target/idl/conditional_vault.json");
